build: `readlink -f` is not compatible

+++ distclean.sh
@@ -3,12 +3,9 @@
 set -eu
 
 curdir="$(dirname "$0")"
-if which readlink >/dev/null 2>&1 ; then
-    curdir="$(readlink -f "$curdir")"
-fi
-
 cd "$curdir"
-if [ ! -x distclean.sh ] || [ ! -f OMakeIncludes ]; then
+
+if [ ! -x distclean.sh ] || [ ! -f OMakeIncludes ] || [ ! -d src ] ; then
     exit 1
 fi
 omake distclean
@@ -16,5 +13,7 @@ find . -type f -name '*.omc' -exec rm {} \+
 find . -type f -name '.omakedb*' -exec rm {} \+
 
 if [ -d .git ]; then
-    git clean -dxf
+    if which git >/dev/null 2>&1; then
+        git clean -dxf
+    fi
 fi
